Moved PruningOperationsMetatypeRegistry to common part
