+++ nncf/common/pruning/model_analysis.py
@@ -19,7 +19,7 @@ from nncf.common.graph.graph import NNCFGraph
 from nncf.common.graph.graph import NNCFNode
 from nncf.common.pruning.export_helpers import DefaultMetaOp
 from nncf.common.pruning.utils import find_next_nodes_not_of_types
-from nncf.pruning.export_utils import PruningOperationsMetatypeRegistry
+from nncf.common.pruning.utils import PruningOperationsMetatypeRegistry
 
 
 # pylint: disable=protected-access
