+++ nncf/common/pruning/utils.py
@@ -17,6 +17,7 @@ from typing import List, Tuple, Optional
 
 from nncf.common.graph.graph import NNCFGraph
 from nncf.common.graph.graph import NNCFNode
+from nncf.common.utils.registry import Registry
 
 
 def is_grouped_conv(node: NNCFNode) -> bool:
@@ -212,3 +213,40 @@ def get_previous_conv(graph: NNCFGraph, nncf_node: NNCFNode,
     if len(sources) == 1 and sources[0].node_type in pruning_types:
         return sources[0]
     return None
+
+
+class PruningOperationsMetatypeRegistry(Registry):
+    def __init__(self, name):
+        super().__init__(name)
+        self._op_name_to_op_class = {}
+
+    def register(self, name=None):
+        name_ = name
+        super_register = super()._register
+
+        def wrap(obj):
+            cls_name = name_
+            if cls_name is None:
+                cls_name = obj.__name__
+
+            super_register(obj, cls_name)
+            op_names = obj.get_all_op_aliases()
+            for name in op_names:
+                name = self.get_version_agnostic_name(name)
+                if name not in self._op_name_to_op_class:
+                    self._op_name_to_op_class[name] = obj
+                else:
+                    assert self._op_name_to_op_class[name] == obj, \
+                        "Inconsistent operator type registry - single patched op name maps to multiple metatypes!"
+            return obj
+
+        return wrap
+
+    def get_operator_metatype_by_op_name(self, op_name: str):
+        if op_name in self._op_name_to_op_class:
+            return self._op_name_to_op_class[op_name]
+        return None
+
+    @staticmethod
+    def get_version_agnostic_name(name):
+        raise NotImplementedError
