+++ nncf/pruning/export_helpers.py
@@ -15,6 +15,7 @@ import torch
 
 from nncf.common.pruning.export_helpers import DefaultMetaOp
 from nncf.common.pruning.utils import is_grouped_conv
+from nncf.common.pruning.utils import get_sources_of_node
 from nncf.dynamic_graph.graph import PTNNCFGraph
 from nncf.dynamic_graph.graph import PTNNCFNode
 from nncf.dynamic_graph.operator_metatypes import NoopMetatype, HardTanhMetatype, TanhMetatype, RELUMetatype, \
@@ -24,13 +25,13 @@ from nncf.dynamic_graph.operator_metatypes import NoopMetatype, HardTanhMetatype
     ConvTranspose2dMetatype, ConvTranspose3dMetatype, GroupNormMetatype
 from nncf.common.utils.logger import logger as nncf_logger
 from nncf.nncf_network import NNCFNetwork
-from nncf.pruning.export_utils import PruningOperationsMetatypeRegistry, identity_mask_propagation, get_input_masks, \
+from nncf.pruning.export_utils import PTPruningOperationsMetatypeRegistry
+from nncf.pruning.export_utils import identity_mask_propagation, get_input_masks, \
     fill_input_masks
-from nncf.common.pruning.utils import get_sources_of_node
 from nncf.layers import NNCF_WRAPPED_USER_MODULES_DICT
 from nncf.pruning.utils import pt_is_depthwise_conv
 
-PT_PRUNING_OPERATOR_METATYPES = PruningOperationsMetatypeRegistry("operator_metatypes")
+PT_PRUNING_OPERATOR_METATYPES = PTPruningOperationsMetatypeRegistry("operator_metatypes")
 
 
 # pylint: disable=protected-access
